xpi/chrome/content/test/test_Database.html

<html>
<head>
	<script type="text/javascript" src="library/MochiKit.js"></script>
	<script type="text/javascript" src="library/SimpleTest.js"></script>
	<script type="text/javascript" src="test.js"></script>
	<link rel="stylesheet" type="text/css" href="library/test.css">
	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
</head>
<body>

<pre id="test">
</pre>
<script type="text/javascript" src="../library/00_MochiKit.js"></script>
<script type="text/javascript" src="../library/00_prototype.js"></script>
<script type="text/javascript" src="../library/00_component.js"></script>
<script type="text/javascript" src="../library/01_utility.js"></script>
<script type="text/javascript" src="../library/10_Progress.js"></script>
<script type="text/javascript" src="../library/10_Database.js"></script>
<script type="text/javascript" src="../library/20_Tombloo.js"></script>
<script type="text/javascript" src="../library/20_Tumblr.js"></script>
<script type="text/javascript" src="../library/30_Tombloo.Service.js"></script>

<script type="text/javascript">
try {
	// chrome://tombloo/content/test/test_Database.html
	
	var file = copy(
		getTestFile('tombloo_test.sqlite'), 
		getTestFile('tombloo_test.temp.sqlite'));
	
	// PRAGMA user_version
	var db = new Database(file);
	is(db.version, 0);
	db.version = 1;
	is(db.version, 1);
	
	// vacuum
	db.vacuum();
	
	var def = {
		name : 'video',
		fields : {
			id      : 'INTEGER PRIMARY KEY',
			host    : 'TEXT',
		}
	}
	
	is(Entity.createInsertSQL(def), 
		'INSERT INTO video ( id, host ) VALUES ( :id, :host )');
	is(Entity.createUpdateSQL(def),
		'UPDATE video SET host=:host WHERE id = :id');
	is(Entity.createInitializeSQL(def),
		'CREATE TABLE IF NOT EXISTS video ( id INTEGER PRIMARY KEY, host TEXT )');
	
	var def = {
		name : 'video',
		fields : {
			id      : 'INTEGER PRIMARY KEY',
			host    : 'TEXT',
			date    : 'TIMESTAMP',
			capTion : 'TEXT',
			source  : 'TEXT',
			player  : 'TEXT',
			tags    : 'LIST',
		}
	}
	
	var Video = Entity(def);
	Video.db = db;
	Video.initialize();
	Video.deinitialize();
	Video.initialize();
	
	ok( db.tableExists('video'));
	ok(!db.tableExists('no_such_table'));
	
	Video.deleteAll();
	is(Video.findAll().length, 0, '全削除されたか');
	
	var video = new Video({
		id : 1,
		host    : 'HOST_1',
		date    : new Date(2007, 6, 20),
		capTion : 'CAPTION',
		tags    : [1,2,3],
	});
	video.save();
	is(Video.findAll().length, 1, 'save');
	
	var video = Video.findByHost('HOST_1')[0];
	is(video.date.getDate(), 20, '日付の保存と復元が行えるか');
	is(video.tags[2], 3, 'リストの保存と復元が行えるか');
	
	
	var video = new Video({
		id : 2,
		host : 'HOST_2',
		date : new Date(2007, 1, 1),
		capTion : 'CAPTION',
	});
	video.save();
	is(Video.findAll().length, 2, 'save');
	
	
	is(Video.findByHost('HOST_1').length, 1);
	is(Video.findByHostAndCapTion(['HOST_1', 'CAPTION']).length, 1);
	
	is(Video.findByCapTion({
		capTion:'CAPTION', 
		limit:1, 
		offset:1,
		order:'id DESC'})[0].host, 'HOST_1');
	
	is(Video.countByHost('HOST_1'), 1);
	is(Video.countByHost({host:'HOST_1'}), 1);
	
	
	// トランザクション
	db.transaction(function(){
		var video = new Video({id : 3});
		video.save();
		
		// エラーを起こしロールバックさせる
		throw 'ERROR';
	}).addErrback(function(err){
		is(err.message, 'ERROR');
	});
	is(Video.findById(3).length, 0, 'ロールバックされたか');

	db.transaction(function(){
		var video = new Video({id : 3});
		video.save();
		
		return 'OK';
	}).addCallback(function(res){
		is(res, 'OK', 'トランザクション結果値が返されるか');
	});
	is(Video.findById(3).length, 1, 'コミットされたか');
	
	
	// トランザクション(個別メソッド)
	db.beginTransaction();
	db.beginTransaction();
	ok(true, 'beginTransaction: トランザクション内で呼び出してもエラーが起きないか');
	
	db.commitTransaction();
	db.commitTransaction();
	ok(true, 'commitTransaction: トランザクション外で呼び出してもエラーが起きないか');
	
	db.rollbackTransaction();
	ok(true, 'rollbackTransaction: トランザクション外で呼び出してもエラーが起きないか');
	
} catch (err) {
	var s = [];
	s.push('TEST SUITE FAILURE!');
	for(var prop in err)
		s.push(prop + ':  ' + err[prop]);
	ok(false, s.join('\n'));
	
} finally{
	db.close();
}
</script>

</body>
</html>
